Data Analytics/requirements.txt
pandasai 
openpyxl 
langchain_groq 
langchain_community
chainlit
python-dotenv